% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2.R
\name{d2}
\alias{d2}
\title{Function to calculate second derivatives from reflectance spectra}
\usage{
d2(wl, w, knots = 0.08, reflectance = FALSE, smooth_reflectance)
}
\arguments{
\item{w}{reflectance values}

\item{knots}{smoothing parameter for the loess function (default = 0.08)}

\item{reflectance}{optional parameter to preview the effect of smoothing on the reflectance data (default=FALSE)}

\item{smooth_reflectance}{(default = 0.08)}
}
\value{
a list with two dataframes: one with the original data and another with the derivative values
}
\description{
Function to calculate second derivatives from reflectance spectra
}
\keyword{external}
