% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{eff}
\alias{eff}
\title{calculate PSII quantum efficiency}
\usage{
eff(f, fm)
}
\arguments{
\item{f}{fluorescence yield}

\item{fm}{maximum fluorescence yield}
}
\value{
numeric vector with the PSII quantum efficiency values
}
\description{
calculate PSII quantum efficiency
}
\keyword{external}
