% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{ETR_eq1}
\alias{ETR_eq1}
\title{ETR calculation from equation 1 Schuback et al. 2021}
\usage{
ETR_eq1(light, sigmaPSII, Fo, Fm, Fp, Fmp)
}
\arguments{
\item{light}{light values (numerical vector)}

\item{sigmaPSII}{Sigma PSII (dark adapted) (single value)}

\item{Fo}{minimum fluorescence yield in dark adapted state (single value)}

\item{Fm}{maximum fluorescence yield in dark adapted state (single value)}

\item{Fp}{F', minimum fluorescence yield in light adapted state (numerical vector)}

\item{Fmp}{maximum fluorescence yield in light adapted state (numerical vector)}
}
\value{
numeric vector with the ETR values
}
\description{
JPII = E x sigma PII x (Fv/Fm)−1 x ( Fq'/Fm')
}
\keyword{external}
