% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{ETR_eq2}
\alias{ETR_eq2}
\title{ETR calculation from equation 2 Schuback et al. 2021}
\usage{
ETR_eq2(light, sigmaPSIIp, Fop, Fmp, Fp)
}
\arguments{
\item{light}{light values (numerical vector)}

\item{sigmaPSIIp}{Sigma' PSII (light adapted) (numerical vector)}

\item{Fop}{minimum fluorescence yield in light adapted state with open reaction centers (numerical vector)}

\item{Fmp}{maximum fluorescence yield in light adapted state (numerical vector)}

\item{Fp}{F', minimum fluorescence yield in light adapted state (numerical vector)}
}
\value{
numeric vector with the ETR values
}
\description{
JSPII = E x Sigma PSII' x (Fq'/Fv'), Fq'/Fv' , calculated as (Fm' -F')/(Fm' - Fo')
}
\keyword{external}
