% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{ETR_eq3}
\alias{ETR_eq3}
\title{ETR calculation from equation 3 Schuback et al. 2021}
\usage{
ETR_eq3(light, sigmaPSIIp, tau1)
}
\arguments{
\item{light}{light values (numerical vector)}

\item{sigmaPSIIp}{Sigma' PSII (light adapted) (numerical vector)}

\item{tau1}{tau1 reopening rate (numerical vector)}
}
\value{
numeric vector with the ETR values
}
\description{
JPSII = E x Sigma PSII' x (1 / (1 + (Sigma PSII' x E x tau1 )))
}
\keyword{external}
