% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{ETR_eq5}
\alias{ETR_eq5}
\title{ETR calculation from equation 5 Schuback et al. 2021}
\usage{
ETR_eq5(light, Fmp, Fo, Fm, Fp, Ka)
}
\arguments{
\item{light}{light values (numerical vector)}

\item{Fmp}{maximum fluorescence yield in light adapted state (numerical vector)}

\item{Fo}{minimum fluorescence yield in dark adapted state (single value)}

\item{Fm}{maximum fluorescence yield in dark adapted state (single value)}

\item{Fp}{F', minimum fluorescence yield in light adapted state (numerical vector)}

\item{Ka}{machine specific calibration factor (single value)}
}
\value{
numeric vector with the ETR values
}
\description{
JVII = E x Fq'/Fm' x (Ka x (Fm x Fo) / Fv)
}
\keyword{external}
