% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npq_models.R
\name{fit_npq}
\alias{fit_npq}
\title{Function to fit NPQ model}
\usage{
fit_npq(light, npq, starting_values = NULL, num_obs = NULL,
  plots = TRUE)
}
\arguments{
\item{light}{Light values (numeric vector)}

\item{npq}{npq values}

\item{starting_values}{the starting values of the parameters to fit (list).}

\item{num_obs}{Number of observations to fit the model (optional), if nothing is defined it will use the number of light steps}

\item{plots}{Turns on and off plotting of data and fitted curves (TRUE, FALSE). Default is TRUE.}
}
\value{
The function returns a list with: 1- fitted parameters and respective; 2 - modeled light levels (numeric vector); 3- predicted NPQ values (numeric vector).
}
\description{
Function to fit NPQ model using the MODEL
}
\keyword{external}
