% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npq_models.R
\name{fit_npq_2011}
\alias{fit_npq_2011}
\title{Function to fit NPQ model}
\usage{
fit_npq_2011(light, npq, starting_values = NULL, num_obs = NULL, plots = TRUE)
}
\arguments{
\item{light}{Light values (numeric vector)}

\item{npq}{npq values}

\item{starting_values}{The starting values of the parameters to fit (list).}

\item{num_obs}{Number of observations to fit the model (optional), if nothing is defined it will use the number of light steps}

\item{plots}{Turns on and off plotting of data and fitted curves (TRUE, FALSE). Default is TRUE.}
}
\value{
The function returns a list with: 1- par and NPQ values of the fitted model; 2 - fitted parameters.
}
\description{
Function to fit NPQ model using the Serodio & Lavaud 2011
}
\keyword{external}
