% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curve.R
\name{gompertz}
\alias{gompertz}
\title{Fit growth curve model to cell counts}
\usage{
gompertz(days, cells, A, umax, l)
}
\arguments{
\item{days}{numerical vector with the time information}

\item{cells}{numerical vector with the cell growth per time information}

\item{A}{number with an estimation of "A" starting value for the model fit}

\item{umax}{number with an estimation of "umax" starting value for the model fit}

\item{l}{number with an estimation of "l" starting value for the model fit}
}
\value{
A list with 3 dataframes: parameters, model and original. The dataframe "parameters" contains the
parameters estimated by the model; the dataframe "model" contains the growth data estimated by the model;
and the dataframe "original" contains the values used to fit the model.
}
\description{
Fits growth curve model (Gompertz, 1825) to cell per day and returns
the parameters: A, umax and l. Where A is the maximum concentration of the biomass (e.g. cells . ml-1),
μmax is maximum growth rate (e.g. cells . day-1) and l is the latency time (e.g. days)
}
\keyword{external}
