% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{integral}
\alias{integral}
\title{Function to calculate the integral of derivative peaks}
\usage{
integral(x, y)
}
\arguments{
\item{x}{x variable, numerical vector}

\item{y}{y variable, numerical vector}
}
\value{
a number with the integral value
}
\description{
Function to calculate the integral of derivative peaks
}
\keyword{external}
