% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_curvesO2.R
\name{light_curvesO2}
\alias{light_curvesO2}
\title{Function to fit PI models to O2 data from the FireSting (Pyroscience)}
\usage{
light_curvesO2(a = "./", light_time)
}
\arguments{
\item{a}{File with the O2 data}

\item{light_time}{Dataframe with one column with light data, one column with the starting times and one column with the finishing times}
}
\description{
Function to fit PI models to O2 data from the FireSting (Pyroscience)
}
\keyword{external}
