% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpbom.R
\name{mpbom}
\alias{mpbom}
\title{Function to apply the MPBOM}
\usage{
mpbom(wl, reflectance, xmin, xmax)
}
\arguments{
\item{wl}{vector with the wavelength values}

\item{reflectance}{vector with the reflectance values}

\item{xmin}{value with the minimum wl}

\item{xmax}{value with the maximum wl}
}
\value{
a list with the following objects:
}
\description{
Applies the MPBOM and calculates the associated indices. It's based on
the paper: "Launeau, Patrick; Méléder, Vona; Verpoorter, Charles; Barillé, Laurent; Kazemipour-Ricci, Farzaneh; Giraud, Manuel; Jesus, Bruno; Le Menn, Erwan. 2018. "Microphytobenthos Biomass and Diversity Mapping at Different Spatial Scales with a Hyperspectral Optical Model" Remote Sens. 10, no. 5: 716. https://doi.org/10.3390/rs10050716 "
}
\keyword{external}
