% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot.R
\name{my_plot}
\alias{my_plot}
\title{plot with standard deviation bars}
\usage{
my_plot(
  x,
  y,
  sd = NULL,
  sd_x = NULL,
  overlay = NULL,
  ylab = NULL,
  xlab = NULL,
  ylim = NULL,
  xlim = NULL,
  pch = NULL,
  bg = NULL,
  type = NULL,
  xaxt = NULL,
  yaxt = NULL,
  lty = NULL,
  col = NULL,
  col_sd = NULL,
  cex = NULL,
  cex.axis = NULL,
  cex.lab = NULL,
  log = NULL
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{sd}{numerical vector with the dispersion measurement to plot}

\item{sd_x}{numerical vector with the dispersion measurement in the x-axis to plot}

\item{overlay}{logical, if TRUE then the function "points" is used instead of "plot"}

\item{ylab}{ylab}

\item{xlab}{xlab}

\item{ylim}{ylim}

\item{xlim}{xlim}

\item{pch}{type of symbol}

\item{bg}{color of the symbol background}

\item{type}{type of plot}

\item{xaxt}{xaxt}

\item{yaxt}{yaxt}

\item{lty}{type of line}

\item{col}{color of the symbol}

\item{col_sd}{color of the sd bars}

\item{cex}{cex}

\item{cex.axis}{cex.axis}

\item{cex.lab}{cex.lab}

\item{log}{log}
}
\value{
plots a x,y plot with sd bars
}
\description{
a customized plot function to plot x,y and SD bars. With the parameter "overlay = TRUE" it uses the function points to overall a data series.
}
\keyword{external}
