% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectance_indices.R
\name{ndvi}
\alias{ndvi}
\title{Function to calculate normalized difference vegetation index (NDVI)}
\usage{
ndvi(red, nir)
}
\arguments{
\item{red}{path to the file to import (text vector)}

\item{nir}{parameter to activate plotting (default = TRUE)}
}
\value{
a list with the following objects: filename- name of the imported file, metadata- all the metadata in the file header, datetime- time and date of the aquisition, all_data- a dataframe with the wavelengths and respective reflectance values.
}
\description{
Function to calculate NDVI
}
\keyword{external}
