% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{npq}
\alias{npq}
\title{calculate NPQ}
\usage{
npq(fm, fmp)
}
\arguments{
\item{fm}{maximum fluorescence yield in dark adapted state}

\item{fmp}{maximum fluorescence yield in dark adapted state}
}
\value{
numeric vector with the NPQ values
}
\description{
calculate NPQ
}
\keyword{external}
