% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_file_irga.R
\name{parse_file_irga}
\alias{parse_file_irga}
\title{Function to import IRGA files}
\usage{
parse_file_irga(file_path)
}
\arguments{
\item{file_path}{Path to the file}
}
\value{
The function returns a dataframe with only the columns used in the calculations.
}
\description{
Function to import and parse files created by the EGM-4 IRGA machine
}
\keyword{external}
