% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_asd.R
\name{read_asd}
\alias{read_asd}
\title{Function to read reflectance spectra from the ASD machine (text format)}
\usage{
read_asd(filename = "./", plot = TRUE, indices = FALSE)
}
\arguments{
\item{filename}{path to the file to import (text vector)}

\item{plot}{parameter to activate plotting (default = TRUE)}

\item{indices}{parameter to activating calculating indices (default = FALSE)}
}
\value{
a list with the following objects: filename- name of the imported file, metadata- all the metadata in the file header, datetime- time and date of the aquisition, all_data- a dataframe with the wavelengths and respective reflectance values.
}
\description{
Function to read reflectance spectra from the ASD machine (text format)
}
\keyword{external}
