% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlc_preview_plots.R
\name{rlc_analysis}
\alias{rlc_analysis}
\title{function to fit several PI models}
\usage{
rlc_analysis(
  x,
  starting_values = list(alpha = 0.5, Eopt = 500, etrmax = 100),
  model = NULL
)
}
\arguments{
\item{x}{rlc object from the process_psi() function}
}
\value{
the fitted values for several PI models
}
\description{
Preview of RLC curves and associated parameters
}
\keyword{external}
