% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlc_fit.R
\name{rlc_fit}
\alias{rlc_fit}
\title{Function to lunch a shiny App that fits RLC models to fluorescence data}
\usage{
rlc_fit()
}
\value{
It does not return any R object
}
\description{
Function to lunch a shiny App that fits RLC models to fluorescence data
}
\keyword{external}
