% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{sigma_Fvp_Fmp}
\alias{sigma_Fvp_Fmp}
\title{function to plot Sigma'/Sigma vs Fv'/Fm'/Fv/Fm}
\usage{
sigma_Fvp_Fmp(
  sigmaPSII,
  sigmaPSIIp,
  Fo,
  Fop,
  Fm,
  Fp,
  Fmp,
  overlay = FALSE,
  pch = NULL,
  bg = NULL
)
}
\arguments{
\item{sigmaPSII}{sigmaPSII}

\item{sigmaPSIIp}{sigmaPSII prime}

\item{Fo}{minimum fluorescence yield}

\item{Fop}{minimum fluorescence at light}

\item{Fm}{maximum fluorescence yield}

\item{Fp}{fluorescence yield at light}

\item{Fmp}{maximum fluorescence yield at light}

\item{overlay}{FALSE}

\item{pch}{NULL}

\item{bg}{NULL}
}
\value{
a plot of Sigma'/Sigma vs Fv'/Fm'/Fv/Fm
}
\description{
function to plot Sigma'/Sigma vs Fv'/Fm'/Fv/Fm
}
\keyword{external}
