% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{yno}
\alias{yno}
\title{calculate YNO}
\usage{
yno(f, fm)
}
\arguments{
\item{f}{fluorescence yield}

\item{fm}{maximum fluorescence yield in dark adapted state}
}
\value{
numeric vector with the YNO values
}
\description{
calculate YNO
}
\keyword{external}
