% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{ynpq}
\alias{ynpq}
\title{calculate YNPQ}
\usage{
ynpq(f, fm, fmp)
}
\arguments{
\item{f}{fluorescence yield}

\item{fm}{maximum fluorescence yield in dark adapted state}

\item{fmp}{maximum fluorescence yield in dark adapted state}
}
\value{
numeric vector with the YNPQ values
}
\description{
calculate YNPQ
}
\keyword{external}
