% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot.R
\name{my_plot}
\alias{my_plot}
\title{plot with standard deviation bars}
\usage{
my_plot(
  x,
  y,
  sd = NULL,
  overlay = NULL,
  ylab = NULL,
  xlab = NULL,
  ylim = NULL,
  xlim = NULL,
  pch = NULL,
  bg = NULL,
  type = NULL,
  xaxt = NULL,
  yaxt = NULL
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{sd}{numerical vector with the dispersion measurement to plot}

\item{overlay}{logical, if TRUE then the function "points" is used instead of "plot"}

\item{ylab}{ylab}

\item{xlab}{xlab}

\item{ylim}{ylim}

\item{xlim}{xlim}

\item{pch}{type of symbol}

\item{bg}{color of the symbol}

\item{type}{type of plot}

\item{xaxt}{xaxt}

\item{yaxt}{yaxt}
}
\value{
plots a x,y plot with sd bars
}
\description{
a customized plot function to plot x,y and SD bars. With the parameter "overlay = TRUE" it uses the function points to overall a data series.
}
\keyword{external}
