% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fluorescence.R
\name{sigma_Fvp_Fmp}
\alias{sigma_Fvp_Fmp}
\title{}
\usage{
sigma_Fvp_Fmp(
  sigmaPSII,
  sigmaPSIIp,
  Fo,
  Fop,
  Fm,
  Fp,
  Fmp,
  overlay = FALSE,
  pch = NULL,
  bg = NULL
)
}
\arguments{
\item{f}{fluorescence yield}

\item{fm}{maximum fluorescence yield}
}
\value{

}
\description{

}
\keyword{external}
